fixed more stuff

+++ app/src/main/java/com/cromiumapps/gravwar/Move.java
@@ -5,14 +5,14 @@ import com.cromiumapps.gravwar.planets.Planet.PlanetType;
 import com.cromiumapps.gravwar.exceptions.InvalidMoveException;
 
 public class Move{	
-	public float missilesToFireAmmount;
+	public int missilesToFireAmmount;
 	public boolean isAiMove;
 	public PlanetType fromPlanetType;
 	public PlanetType toPlanetType;
-	public float fromPlanetId;
-	public float toPlanetId;
+	public int fromPlanetId;
+	public int toPlanetId;
 	
-	public Move(float missilesToFireAmmount, Planet from, Planet to, boolean isAiMove) throws InvalidMoveException
+	public Move(int missilesToFireAmmount, Planet from, Planet to, boolean isAiMove) throws InvalidMoveException
 	{
 		if(from.getId() == to.getId()) throw new InvalidMoveException("fromid and toid are identical");
 		if(missilesToFireAmmount <= 0) throw new InvalidMoveException("missiles to fire are <= 0");
